Fix a problem in the ".recover" command allowing a circular loop of b-tree pages in a database file to cause an infinite loop.
FossilOrigin-Name: 8d2a062eb8a3e6fdc6a61b571c8da0070382bf208c53e797151eac8679c975a1

+++ ext/misc/dbdata.c
@@ -301,7 +301,10 @@ static unsigned int get_uint16(unsigned char *a){
   return (a[0]<<8)|a[1];
 }
 static unsigned int get_uint32(unsigned char *a){
-  return (a[0]<<24)|(a[1]<<16)|(a[2]<<8)|a[3];
+  return ((unsigned int)a[0]<<24)
+       | ((unsigned int)a[1]<<16)
+       | ((unsigned int)a[2]<<8)
+       | ((unsigned int)a[3]);
 }
 
 /*
